+++ panoptes/utils/tests/images/test_fits_utils.py
@@ -1,10 +1,11 @@
 import os
-import pytest
-import subprocess
 import shutil
+import subprocess
+from contextlib import suppress
 
-from astropy.io.fits import Header
+import pytest
 from astropy import units as u
+from astropy.io.fits import Header
 
 from panoptes.utils.images import fits as fits_utils
 
@@ -34,6 +35,29 @@ def test_fpack(solved_fits_file):
     assert os.stat(compressed).st_size == info.st_size
 
     os.remove(copy_file)
+
+
+def test_no_overwrite_fpack(solved_fits_file):
+    new_file = solved_fits_file.replace('solved', 'solved_copy')
+    copy_file = shutil.copyfile(solved_fits_file, new_file)
+
+    # Unpack the file. This removes the packed version.
+    uncompressed = fits_utils.funpack(copy_file)
+
+    # Copy file again so now the packed version exists alongside unpacked.
+    copy_file = shutil.copyfile(solved_fits_file, new_file)
+
+    # Deny overwriting gives error.
+    with pytest.raises(FileExistsError):
+        _ = fits_utils.fpack(uncompressed, overwrite=False)
+
+    # Default is overwrite=True.
+    compressed = fits_utils.fpack(uncompressed)
+
+    # Cleanup test.
+    for file in [copy_file, uncompressed, compressed]:
+        with suppress(FileNotFoundError):
+            os.remove(file)
 
 
 def test_getheader(solved_fits_file):
@@ -67,6 +91,10 @@ def test_solve_field_unsolved(unsolved_fits_file):
     wcs_info = fits_utils.get_wcsinfo(unsolved_fits_file.replace('.fits', '.new'))
     assert 'crpix0' in wcs_info
     assert wcs_info['crpix0'] == 350.5 * u.pixel
+
+    for ext in ['.wcs', '.new']:
+        with suppress(FileNotFoundError):
+            os.remove(unsolved_fits_file.replace('.fits', ext))
 
 
 def test_get_solve_field_solved(solved_fits_file):