+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 
 ## [0.2.dev11]
 
+### Added
+
+* Data
+    * Added basic data access components for getting observation and image metadata. (#178, #181)
+    * Added a `search_observations` function for searching by various criteria. (#181)
+        * Uses anonymous credentials to connect to firestore.
+        * Added a basic notebook demonstrating features.
+    * Adding `holoviews` and `hvplot` as required dependencies.
+    
+
 ### Bug fixes
 
 * FITS Utils fixes:
@@ -28,6 +38,9 @@ All notable changes to this project will be documented in this file.
     * Changed default options in `get_solve_field` to use `scale-low` and `scale-high` instead of `radius` (which
     requires an `ra` and `dec`). (#180)
 * Changed `bin/panoptes-dev` -> `bin/panoptes-develop` for naming consistency. (#175)
+* Data
+    * **BREAKING** The `panoptes.utils.data.py` has moved into the `panoptes.utils.data` namespace with the relevant existing `Downloader` class placed in the `assets.py` module. (#181)
+    * Changed the `get_data` (and images and observations equivalent) to `get_metadata`. (#181)
 
 ### Removed